refactor: Remove flavor nested enum and add fatContent property in Milk
